requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-migrate==3.0.1
python-dotenv==0.18.0
pycodestyle==2.7.0
psycopg2==2.9.1
python-jose==3.3.0